+++ client/bucket.go
@@ -29,3 +29,28 @@ func (c *CSClient) ListBuckets(ctx context.Context, authToken string) (*ListBuck
 	defer httpResp.Body.Close()
 	return &resp, nil
 }
+
+func (c *CSClient) ReadBucketDataset(
+	ctx context.Context,
+	authToken,
+	bucketName string,
+) (*ReadBucketDatasetResp, error) {
+	var datasetResp ReadBucketDatasetResp
+
+	httpResp, err := c.createAndSendReq(ctx, ClientRequest{
+		Url:         fmt.Sprintf("%s/Bucket/dataset/name/%s?state=true", c.Config.URL, bucketName),
+		RequestType: GET,
+		AuthToken:   authToken,
+	})
+
+	if err != nil {
+		return nil, fmt.Errorf("Read Bucket Metadata Failure => %s", err)
+	}
+
+	if err := c.unmarshalJSONBody(httpResp.Body, &datasetResp); err != nil {
+		return nil, err
+	}
+
+	defer httpResp.Body.Close()
+	return &datasetResp, nil
+}