+++ provider/resources/objectgroup.go
@@ -8,6 +8,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"reflect"
+	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -105,9 +106,25 @@ func ResourceObjectGroup() *schema.Resource {
 					},
 				},
 			},
-			"live_events": {
+			"live_events_aws": {
 				Type:     schema.TypeString,
 				Optional: true,
+			},
+			"live_events_gcp": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"project_id": {
+							Type:     schema.TypeString,
+							Required: true,
+						},
+						"subscription_id": {
+							Type:     schema.TypeString,
+							Required: true,
+						},
+					},
+				},
 			},
 			"index_retention": {
 				Type:     schema.TypeSet,
@@ -495,7 +512,6 @@ func resourceObjectGroupCreate(ctx context.Context, data *schema.ResourceData, m
 		Format:            format,
 		IndexRetention:    indexRetention,
 		Filter:            filters,
-		LiveEvents:        data.Get("live_events").(string),
 		PartitionBy:       data.Get("partition_by").(string),
 		TargetActiveIndex: data.Get("target_active_index").(int),
 		Options:           options,
@@ -504,6 +520,21 @@ func resourceObjectGroupCreate(ctx context.Context, data *schema.ResourceData, m
 			Column: 0,
 		},
 		Realtime: false,
+	}
+
+	liveEventsAws := data.Get("live_events_aws").(string)
+	liveEventsGcp := data.Get("live_events_gcp").(*schema.Set).List()
+	if liveEventsAws != "" && len(liveEventsGcp) > 0 {
+		err := "Live Events found defined for both AWS and GCP, please ensure you configure only one for your cluster type"
+		return diag.Errorf(err)
+	} else if liveEventsAws != "" {
+		createObjectGroupRequest.LiveEventsAws = liveEventsAws
+	} else if len(liveEventsGcp) > 0 {
+		liveEventsMap := liveEventsGcp[0].(map[string]interface{})
+		createObjectGroupRequest.LiveEventsGcp = &client.LiveEventsGcp{
+			ProjectId:      liveEventsMap["project_id"].(string),
+			SubscriptionId: liveEventsMap["subscription_id"].(string),
+		}
 	}
 
 	if err := c.CreateObjectGroup(ctx, createObjectGroupRequest); err != nil {
@@ -701,20 +732,42 @@ func ResourceObjectGroupRead(ctx context.Context, data *schema.ResourceData, met
 	}
 
 	if resp.Options != nil {
-		options := map[string]interface{}{
-			"compression": resp.Compression,
+		var options map[string]interface{}
+		optionsList := data.Get("options").(*schema.Set).List()
+		optionsMap := optionsList[0].(map[string]interface{})
+		compression := optionsMap["compression"].(string)
+
+		if strings.EqualFold(resp.Options.Compression, compression) {
+			options = map[string]interface{}{"compression": compression}
+		} else {
+			options = map[string]interface{}{"compression": resp.Options.Compression}
 		}
 
-		if len(resp.Options.ColTypes) > 0 {
+		colTypesString := optionsMap["col_types"].(string)
+		if colTypesString != "" {
+			options["col_types"] = colTypesString
+		} else if len(resp.Options.ColTypes) > 0 {
 			colTypes, _ := json.Marshal(resp.Options.ColTypes)
 			colTypesJson, _ := structure.NormalizeJsonString(string(colTypes))
 			options["col_types"] = colTypesJson
 		}
 
-		if len(resp.Options.ColSelection) > 0 {
+		colSelectionString := optionsMap["col_selection"].(string)
+		if colSelectionString != "" {
+			options["col_selection"] = colSelectionString
+		} else if len(resp.Options.ColSelection) > 0 {
 			colSelect, _ := json.Marshal(resp.Options.ColSelection)
 			colSelectJson, _ := structure.NormalizeJsonString(string(colSelect))
 			options["col_selection"] = colSelectJson
+		}
+
+		colRenamesString := optionsMap["col_renames"].(string)
+		if colRenamesString != "" {
+			options["col_renames"] = colRenamesString
+		} else if len(resp.Options.ColRenames) > 0 {
+			colRenames, _ := json.Marshal(resp.Options.ColRenames)
+			colRenamesJson, _ := structure.NormalizeJsonString(string(colRenames))
+			options["col_renames"] = colRenamesJson
 		}
 
 		err = data.Set("options", []interface{}{options})